created Category

+++ src/main/java/com/example/OrderCoffeeBE/Service/ProductServiceImpl.java
@@ -1,14 +1,16 @@
 package com.example.OrderCoffeeBE.Service;
 
+import com.example.OrderCoffeeBE.Entity.Category;
 import com.example.OrderCoffeeBE.Entity.Product;
 import com.example.OrderCoffeeBE.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service("productService")
-public class ProductServiceImpl implements IProductService {
+public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductRepository productRepository;
@@ -23,17 +25,29 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public void CreateProduct(Product product) {
-        productRepository.save(product);
+    public Optional<Product> findByName(String name) {
+        return productRepository.findByName(name);
     }
 
     @Override
-    public void UpdateProduct(Product product) {
-        productRepository.save(product);
+    public Product CreateProduct(Product product) {
+        return productRepository.save(product);
     }
 
     @Override
-    public void DeleteProduct(int id) {
-        productRepository.deleteById(id);
+    public Product UpdateProduct(Product product) {
+        return productRepository.save(product);
     }
+
+    @Override
+    public Product DeleteProduct(int id) {
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isPresent()) {
+            productRepository.deleteById(id);
+            return product.get();
+        }
+        return null;
+    }
+
+
 }
